requirements.txt
flask
google-genai
Pillow
python-dotenv
requests
gunicorn
